Move Constants into DarcLib
