+++ src/Microsoft.DotNet.Darc/src/Darc/Helpers/EncodedFile.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using Microsoft.DotNet.DarcLib;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json.Linq;
 using System;
