+++ src/Microsoft.DotNet.Darc/src/Darc/Program.cs
@@ -14,6 +14,7 @@ using Microsoft.ApplicationInsights.Channel;
 using Microsoft.ApplicationInsights.DataContracts;
 using Microsoft.ApplicationInsights.DependencyCollector;
 using Microsoft.ApplicationInsights.Extensibility;
+using Microsoft.DotNet.DarcLib;
 
 namespace Microsoft.DotNet.Darc
 {
